feat: refactor navigation layout and consolidate settings
